src/requirements.txt
flask
gunicorn
google-auth
requests
firebase-admin
python-dotenv
beautifulsoup4
google-cloud-datastore